Update watsonx-ai dependency to 0.16.0

+++ langchain4j-watsonx/src/test/java/dev/langchain4j/model/watsonx/it/WatsonxModerationModelIT.java
@@ -45,7 +45,7 @@ public class WatsonxModerationModelIT {
         assertEquals(11, metadata.get("end"));
         assertEquals("hap", metadata.get("detection_type"));
         assertEquals("has_HAP", metadata.get("detection"));
-        assertEquals(0.98, (float) metadata.get("score"), 0.01);
+        assertEquals(0.98, (double) metadata.get("score"), 0.01);
     }
 
     @Test
@@ -78,6 +78,6 @@ public class WatsonxModerationModelIT {
         assertEquals(11, metadata.get("end"));
         assertEquals("hap", metadata.get("detection_type"));
         assertEquals("has_HAP", metadata.get("detection"));
-        assertEquals(0.98, (float) metadata.get("score"), 0.01);
+        assertEquals(0.98, (double) metadata.get("score"), 0.01);
     }
 }
